NF: move CountModels where it's used
